requirements.txt
astrapy>=0.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
